+++ tong12580.rpc/rpc.common/src/main/java/io/github/tong12580/common/core/lang/Constants.java
@@ -1,0 +1,20 @@
+package io.github.tong12580.common.core.lang;
+
+/**
+ * <p>Constants</p>
+ * <span></span>
+ *
+ * @author yuTong
+ * @version 1.0
+ * @since 2019/6/23 23:17
+ */
+public interface Constants {
+    /**
+     * 默認超時時間 6秒
+     */
+    int DEFAULT_TIMEOUT = 6000;
+    int CLOSE_THRESHOLD_VALUE = 2;
+    String PING = "Ping";
+    String PONG = "Pong";
+    String RPC_SERVER = "rpc_server";
+}
