+++ tong12580.rpc/rpc.common/src/main/java/io/github/tong12580/common/core/lang/DefaultThreadFactory.java
@@ -1,0 +1,43 @@
+package io.github.tong12580.common.core.lang;
+
+import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.atomic.AtomicInteger;
+
+/**
+ * 默认线程配置
+ * DefaultThreadFactory
+ *
+ * @author yuTong
+ * @version 1.0
+ * @since 2019-7-11 17:20
+ */
+public class DefaultThreadFactory implements ThreadFactory {
+
+    private static final AtomicInteger POOL_NUMBER = new AtomicInteger(1);
+    private final ThreadGroup group;
+    private final AtomicInteger threadNumber = new AtomicInteger(1);
+    private final String namePrefix;
+
+    public DefaultThreadFactory(String threadName) {
+        SecurityManager s = System.getSecurityManager();
+        group = (s != null) ? s.getThreadGroup() :
+                Thread.currentThread().getThreadGroup();
+        namePrefix = threadName + "-" +
+                POOL_NUMBER.getAndIncrement() +
+                "-thread-";
+    }
+
+    @Override
+    public Thread newThread(Runnable r) {
+        Thread t = new Thread(group, r,
+                namePrefix + threadNumber.getAndIncrement(),
+                0);
+        if (t.isDaemon()) {
+            t.setDaemon(false);
+        }
+        if (t.getPriority() != Thread.NORM_PRIORITY) {
+            t.setPriority(Thread.NORM_PRIORITY);
+        }
+        return t;
+    }
+}
