+++ tong12580.rpc/src/main/java/io/github/tong12580/rpc/core/binding/RpcServiceProxy.java
@@ -1,9 +1,5 @@
 package io.github.tong12580.rpc.core.binding;
 
-import io.github.tong12580.rpc.common.Constants;
-import io.github.tong12580.rpc.common.cache.ChannelCache;
-import io.netty.channel.Channel;
-
 import java.io.Serializable;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
@@ -18,13 +14,17 @@ import java.lang.reflect.Method;
  */
 public class RpcServiceProxy<T> implements InvocationHandler, Serializable {
 
+    private Class<T> tClass;
+    private String methodName;
+
+    public RpcServiceProxy(Class<T> className, String methodName) {
+        this.tClass = className;
+        this.methodName = methodName;
+    }
+
     @Override
     public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
-        //TODO 1
-        Channel channel = ChannelCache.get(Constants.RPC_SERVER);
-
-        Object object = method.invoke(this, args);
-        //TODO 2
-        return object;
+        method = tClass.getMethod(methodName);
+        return method.invoke(tClass.newInstance(), args);
     }
 }
