+++ tong12580.rpc/src/main/java/io/github/tong12580/rpc/core/binding/RpcServiceProxyFactory.java
@@ -1,5 +1,7 @@
 package io.github.tong12580.rpc.core.binding;
 
+import java.lang.reflect.Proxy;
+
 /**
  * Rpc服务代理工厂
  * RpcServiceProxyFactory
@@ -10,8 +12,20 @@ package io.github.tong12580.rpc.core.binding;
  */
 public class RpcServiceProxyFactory<T> {
     private final Class<T> mapperInterface;
+    private final String methodName;
 
-    public RpcServiceProxyFactory(Class<T> mapperInterface) {
+    public RpcServiceProxyFactory(Class<T> mapperInterface, String methodName) {
         this.mapperInterface = mapperInterface;
+        this.methodName = methodName;
+    }
+
+    @SuppressWarnings("unchecked")
+    private T newInstance(RpcServiceProxy rpcServiceProxy) {
+        return (T) Proxy.newProxyInstance(mapperInterface.getClassLoader(), new Class[]{mapperInterface}, rpcServiceProxy);
+    }
+
+    public T newInstance() {
+        final RpcServiceProxy<T> rpcServiceProxy = new RpcServiceProxy<>(mapperInterface, methodName);
+        return newInstance(rpcServiceProxy);
     }
 }
